Include drafts in pull request previews
When a pull request is built by netlify, include drafts so that the PR
author can see what the site will look like with their change. Keep in
mind that even though they will see drafts in the PR preview, on the
live site they will be hidden.

You should look at the staging website to see draft documents on the
website.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>