Use drafts branch for WIP content
New content that you are still working on can be submitted as pull
request to the drafts branch. This will get the content up on our
drafts website at https://drafts--cncf-contribute.netlify.app where
you and others can collaborate until the content is ready to go live.

At that point, submit your changes to the main branch and have the TOC
review it before merging.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>