Fix twitter cards
The twitter card links rely on the domain of the website being set so that it can be an absolute URL that includes the host, e.g. https://contribute.cncf.io/featured-image.png instead of /featured-image.png

Netlify doesn't set DEPLOY_PRIME_URL for production deployments so I have set the build script to default to that domain when its unset.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>