Ensure npm dependencies are installed on netlify
This automatically generates our website resources files, and removes them from the repository so that we don't need to check them in anymore. When netlify runs our build, it now runs npm install in the website directory and installs the necessary tools (such as postcss, the tools are listed in the website/package.json file) so that when hugo runs a build, that the resources can be generated.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>