Fix links to sample instructions on templates
When a page (such as templates/contributing.md) links to an image, the image source reference needs to be relative to the final URL of the page. Since the image is in templates/ and the page URL is templates/contributing/, the image ref should be ../sample-instructions.png (going up one level so that we are referencing the image in the template dir).

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>