Fix markdown
I have updated the security guidelines page so that it is rendered on the website. It needed a Hugo frontmatter section, and to be named with a .md file extension. We also needed an _index.md for the new security section that it is contained in.

I also fixed the links on the page so that the footnotes render properly. They are now working hyperlinks, and render in a bracket.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>